+++ internal/services/server.go
@@ -8,6 +8,26 @@ import (
 	"github.com/itaraxa/effectivepancake/internal/models"
 )
 
+// Интерфейс для описания взаимодействия с хранилищем метрик
+type Storager interface {
+	UpdateGauge(metricName string, value float64) error
+	AddCounter(metricName string, value int64) error
+	GetMetrica(metricaType string, metricaName string) (string, error)
+	String() string
+	HTML() string
+}
+
+// Интерфейс для описания взаимодействия с запросом на обновление метрики
+type Querier interface {
+	GetMetricaType() string
+	SetMetricaType(string) (Querier, error)
+	GetMetricName() string
+	SetMetricaName(string) (Querier, error)
+	GetMetricaRawValue() string
+	SetMetricaRawValue(string) (Querier, error)
+	String() string
+}
+
 /*
 Creating new instance of models.Query from request.URL string
 
@@ -20,18 +40,18 @@ Returns:
 	q models.Query: copy of instance models.Query
 	err error: nil or error occurred while parsing the raw string
 */
-func ParseQueryString(raw string) (q models.Query, err error) {
+func ParseQueryString(raw string) (q Querier, err error) {
 	queryString := raw[1:]
 	if len(strings.Split(queryString, `/`)) != 4 {
-		return q, errors.ErrBadRawQuery
+		return nil, errors.ErrBadRawQuery
 	}
-	q.MetricType = strings.Split(queryString, `/`)[1]
-	if q.MetricType != "gauge" && q.MetricType != "counter" {
-		return q, errors.ErrBadType
+	q = models.NewQuery()
+	q, err = q.SetMetricaType(queryString)
+	if err != nil {
+		return q, err
 	}
-	q.MetricName = strings.Split(queryString, `/`)[2]
-	q.MetricRawValue = strings.Split(queryString, `/`)[3]
-
+	// q, _ = q.SetMetricaName(queryString)
+	q, _ = q.SetMetricaRawValue(queryString)
 	return q, nil
 }
